requirements.txt
numpy
pandas
requests
datetime
scikit-learn
xgboost
customtkinter
pillow
matplotlib
seaborn